FPCN/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.4.0=pypi_0
ca-certificates=2023.01.10=haa95532_0
cachetools=5.3.1=pypi_0
certifi=2022.12.7=py39haa95532_0
charset-normalizer=3.0.1=pypi_0
clip=1.0=pypi_0
colorama=0.4.6=pypi_0
cycler=0.11.0=pypi_0
filelock=3.9.0=pypi_0
fonttools=4.38.0=pypi_0
ftfy=6.1.1=pypi_0
google-auth=2.22.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grad-cam=1.4.6=pypi_0
grpcio=1.56.0=pypi_0
huggingface-hub=0.12.1=pypi_0
idna=3.4=pypi_0
imageio=2.26.0=pypi_0
importlib-metadata=6.8.0=pypi_0
joblib=1.2.0=pypi_0
kiwisolver=1.4.4=pypi_0
lazy-loader=0.1=pypi_0
markdown=3.4.3=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.5.3=pypi_0
networkx=3.0=pypi_0
numpy=1.22.3=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=1.1.1t=h2bbff1b_0
packaging=23.0=pypi_0
pillow=9.1.0=pypi_0
pip=22.3.1=py39haa95532_0
protobuf=3.20.3=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyparsing=3.0.9=pypi_0
python=3.9.16=h6244533_0
python-dateutil=2.8.2=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
regex=2022.10.31=pypi_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.20.0=pypi_0
scikit-learn=1.1.1=pypi_0
scipy=1.9.1=pypi_0
setuptools=65.6.3=py39haa95532_0
six=1.16.0=pypi_0
sqlite=3.40.1=h2bbff1b_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorboardx=2.6=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.3.15=pypi_0
tokenizers=0.12.1=pypi_0
torch=1.12.1+cu116=pypi_0
torchcam=0.3.2=pypi_0
torchvision=0.14.1=pypi_0
tqdm=4.64.0=pypi_0
transformers=4.19.2=pypi_0
ttach=0.0.3=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.14=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.6=pypi_0
werkzeug=2.3.6=pypi_0
wheel=0.38.4=py39haa95532_0
wincertstore=0.2=py39haa95532_2
zipp=3.16.0=pypi_0
